requirements.txt
#langchain
#python-dotenv
#ipykernel 
#langchain_community
#Pypdf
#bs4
#arxiv
#pymupdf
#pdfplumber
#wikipedia
#langchain-text-splitters
#langchain-openai
#sentence-transformers
#chromadb
#langchain-openai
#streamlit
#langchain_community
#faiss-cpu
#beautifulsoup4
langchain_groq
langchain
groq
streamlit
python-dotenv
langchain_community
langchain_core